# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Advanced web panel for managing Xray-core servers. Supports VMESS, VLESS, Trojan, Shadowsocks, Hysteria, WireGuard, and more.",
        "title": "3X-UI Panel API",
        "termsOfService": "https://github.com/MHSanaei/3x-ui",
        "contact": {
            "name": "MHSanaei",
            "url": "https://github.com/MHSanaei/3x-ui"
        },
        "license": {
            "name": "GNU GPL v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "3.0.1"
    },
    "host": "localhost:2053",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Login with username and password. Returns a session cookie on success. Supports optional 2FA TOTP code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    },
                    "401": {
                        "description": "Wrong username or password",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/csrf-token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a CSRF token for authenticated SPA clients. Requires login session.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get CSRF token (authenticated)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.LoginForm": {
            "description": "Login form with username, password and optional 2FA code.",
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "twoFactorCode": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Use format: \"Bearer \u003ctoken\u003e\". Get your token from Settings → Security → API Token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}